improved image handling
